+++ spec/dummy/config/routes.rb
@@ -12,6 +12,7 @@ Rails.application.routes.draw do
   get "chat" => "demos#chat"
   get "todos" => "demos#todos"
   get "combobox" => "demos#combobox"
+  get "filter_combobox" => "demos#filter_combobox"
   get "new_order" => "demos#new_order"
   get "order/:id" => "demos#order"
   get "post_preview/:id" => "demos#post_preview"